Adds pokedeck vew and a list into it (Warning :  ugly stuff)
