+++ app/src/main/AndroidManifest.xml
@@ -13,20 +13,18 @@
         android:label="@string/app_name"
         android:supportsRtl="true"
         android:theme="@style/AppTheme">
-        <activity
-            android:name=".activity.LoginActivity"
-            android:label="@string/app_name"></activity>
-        <activity
-            android:name=".activity.MainPageActivity"
-            android:configChanges="orientation|keyboardHidden|screenSize"
-            android:label="@string/title_activity_main_page"
-            android:theme="@style/FullscreenTheme">
+        <activity android:name=".activity.PokedeckActivity">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
-
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
+
+        <activity
+            android:name=".activity.LoginActivity"
+            android:label="@string/app_name" />
+
+
     </application>
 
 </manifest>
